[feat] allow for notifications in validation messages
Validation failures that should result in an email or notification need
to include at least a reason for a failure in order to be published.
Updated the validation message model to allow for this as well as a
system_id if we have it so we can identify the system in the
notification.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>