separate kafka-sasl cdappconfig.json
We don't test a local kafka with sasl, so I moved that config out of the
standard cdappconfig.json so I don't break local dev work. I kept it
around in case we want to figure it out  in the future, and made an
entry in the Makefile to use it if we can get it going.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>