Add Kafka Auth Support (#230)
* update clowder version

* add kafka auth to config

* add kafka auth when available

* update cdappconfig to support newer clowder and sasl

* go fmt kafka.go

* add connect sasl config deployment file

for secure kafka environments

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* refactor kafka auth config

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* Setkey can return an error

As far as I can tell, this isn't an error we need. I wasn't able to get
these values added by just trying to add them to the map. There might be
a better way to do it.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* fix secret names in connect-sasl

This configures the secrets with the proper keys and values based on the
RHOSAK configuration doc

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* Remove improper syntax from kafkaConnect

We dont' need a cert with RHOSAK and we can't use the file syntax in
kafkaconnect. We'll need to pull the secret for the directly from vault
in the saas file

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* Setup connect-rhosak for testing in managed kafka

wipe out everything in connect-rhosak.yml so that all we do is connect
to kafka. This will tell us if our config is going to work as we expect

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>

* separate kafka-sasl cdappconfig.json

We don't test a local kafka with sasl, so I moved that config out of the
standard cdappconfig.json so I don't break local dev work. I kept it
around in case we want to figure it out  in the future, and made an
entry in the Makefile to use it if we can get it going.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>